Supports toggling eks auto mode on and off
The API has a number of constraints that make this a litle difficult,
but it does work.

1.  The node_role_arn cannot be changed without recreating the cluster.
    This is an API limitation, not a terraform provider bug.
    https://docs.aws.amazon.com/eks/latest/APIReference/API_ComputeConfigRequest.html

2.  If node_pools is not empty, then the node_role_arn *must* be provided.

3.  If the node_role_arn is provided, then node_pools must not be empty!

If you enable Auto Mode and set node_pools to a non-empty value, which requires
setting node_role_arn, you cannot later change node_pools to an empty value,
unless you also unset the node_role_arn. But if you unset the node_role_arn,
then the cluster will be recreated due to the first point above.

So, we *can* support disabling auto-mode without recreating a cluster, but only
if node_pools is not empty and if node_role_arn does not change.

Fixes https://github.com/terraform-aws-modules/terraform-aws-eks/issues/3273